My Ibaadah Manager app version 0.1.0
Welcome to Ibaadah app version 0.1.0 - alpha! This is our Minimum Viable Product (MVP), and we're excited to share it with you. Although it's still a work in progress, we believe this version provides a solid foundation for our vision of a comprehensive Islamic application that helps Muslims organize their acts of worship.

We're continuously working to improve the app and add new features, so please bear with us as we make updates and enhancements. We're committed to providing the best experience possible for our users, and we value your feedback and suggestions.

In the meantime, we hope you find the current version of the app helpful in planning and organizing your worship routine. Stay tuned for future updates and new features!

https://docs.google.com/forms/d/e/1FAIpQLSfVA4rxiKoUMQ2vd5bXDjqGvmr0flrf2T3JZw-BonfwzSNDdA/viewform?embedded=true